+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2.0.2
+ - Fixed generate playstore icon wrong location ([#19](https://github.com/mrrhak/icons_launcher/issues/19))
+ - Fixed android adaptive round icon not remove ([#18](https://github.com/mrrhak/icons_launcher/issues/18))
+
 ## 2.0.1
  - Fixed bug with `adaptive_background_color` and `adaptive_round_image` on android thanks to [Carapacik](https://github.com/mrrhak/icons_launcher/pull/17)
  - Improve config validation